refactoring to batch calculation to avoid overflows

+++ src/types/Rewards.sol
@@ -14,7 +14,7 @@ struct Holder {
 /// @dev The curve parameters for reward pool share issuance
 struct CurveParams {
     /// @dev the number of periods for which rewards are paid (acts as the exponent)
-    uint8 numPeriods;
+    uint16 numPeriods;
     /// @dev The base of the exponential formula for reward calculations
     uint16 formulaBase;
     /// @dev the period duration in seconds
@@ -22,7 +22,7 @@ struct CurveParams {
     /// @dev the start timestamp for rewards
     uint48 startTimestamp;
     /// @dev the minimum multiplier for rewards
-    uint16 minMultiplier;
+    uint256 minMultiplier;
 }
 
 /// @dev The slashing parameters for the reward pool. Slashing is a mechanism to to burn shares for a
